# add hash key and early stop mechanism for libtuner