--- a/src/flag_gems/utils/libentry.py
+++ b/src/flag_gems/utils/libentry.py
@@ -28,14 +28,16 @@ version = triton.__version__.split(".")
 major_version, minor_version = eval(version[0]), eval(version[1])
 
 
-def get_kernel_hash(func):
+def get_kernel_hash(func, configs):
     if hasattr(func, "fn"):
         original_func = func.fn
     else:
         original_func = func
 
     source_code = inspect.getsource(original_func)
-    return hashlib.md5(source_code.encode("utf-8")).hexdigest()[:8]
+    config_strs = [str(config) for config in configs]
+    combined_content = f"{source_code}{config_strs}"
+    return hashlib.md5(combined_content.encode("utf-8")).hexdigest()[:8]
 
 
 if major_version == 2:
@@ -203,7 +205,7 @@ class LibTuner(triton.runtime.Autotuner):
         self.keys = key
         self.strategy = strategy
         self.share = share
-        self.kernel_hash = get_kernel_hash(self.base_fn)
+        self.kernel_hash = get_kernel_hash(self.base_fn, self.configs)
         # Use table name with hash instead of hash in key
         self.table_name = (
             f"{share}_{self.kernel_hash}"