@meinie0826 2025-07-01
> 如果您可以演示启用 early_stop 之前和之后 matmul 函数之间的性能比较，那将是有益的。
As I mentioned earlier, by running the script above, the test results before and after opening are as follows:
Before:
```
Adding requests: 100%|███████████████████████████████████████████████████| 4/4 [00:00<00:00, 110.27it/s]
Processed prompts: 100%|█| 4/4 [00:11<00:00,  2.87s/it, est. speed input: 1.92 toks/s, output: 38.20 tok
Time taken for generation: 11.51 seconds
Adding requests: 100%|██████████████████████████████████████████████████| 4/4 [00:00<00:00, 1592.67it/s]
Processed prompts: 100%|█| 4/4 [00:01<00:00,  3.46it/s, est. speed input: 19.05 toks/s, output: 415.69 t
Time taken for generation: 1.16 seconds
```
After:
```
Adding requests: 100%|███████████████████████████████████████████████████| 4/4 [00:00<00:00, 288.53it/s]
Processed prompts: 100%|█| 4/4 [00:08<00:00,  2.13s/it, est. speed input: 2.58 toks/s, output: 51.45 tok
Time taken for generation: 8.53 seconds
Adding requests: 100%|██████████████████████████████████████████████████| 4/4 [00:00<00:00, 1959.27it/s]
Processed prompts: 100%|█| 4/4 [00:01<00:00,  2.40it/s, est. speed input: 13.18 toks/s, output: 287.53 t
Time taken for generation: 1.67 seconds
```
There are two running times before and after, the first time represents the running time without cache, and the second time represents the running time after tuning. In fact, this is related to the two hyperparameters early_stop_threshold and early_stop_min_comfigs. We can also consider adding a flag to choose whether to enable the early stop function. I think this function is meaningful for non long and small tasks.