@meinie0826 2025-07-01
Based on your suggestion, I retested the end-to-end latency
```
Adding requests: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 4/4 [00:00<00:00, 294.44it/s]
Processed prompts: 100%|██████████████████████████████████████████████████████████████| 4/4 [00:08<00:00,  2.12s/it, est. speed input: 2.59 toks/s, output: 51.56 toks/s]
Time taken for generation: 8.51 seconds
Adding requests: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 4/4 [00:00<00:00, 1281.88it/s]
Processed prompts: 100%|████████████████████████████████████████████████████████████| 4/4 [00:01<00:00,  2.41it/s, est. speed input: 13.28 toks/s, output: 289.70 toks/s]
Time taken for generation: 1.66 seconds
```