@sgjzfzzf 2025-07-07
> I have retest on the follow script:
> 
> ```
> import torch
> 
> import flag_gems
> 
> flag_gems.enable()
> 
> 
> def test_mm_configs():
>     device = torch.device("cuda")
>     nk_combinations = [
>         [3584, 3584],
>         [18944, 3584],
>         [3584, 18944],
>         [152064, 3584],
>         [37888, 3584],
>     ]
> 
>     m_values = list(range(1, 1025, 16))  
> 
>     total_times = [0, 0]  # [tuning_time, cached_time]
> 
>     for round_idx in range(6):
>         print("=" * 50)
>         for N, K in nk_combinations:
>             print(f"test N={N}, K={K}")
>             print("-" * 30)
> 
>             for M in m_values:
>                 A = torch.randn(M, K, device=device, dtype=torch.float16)
>                 B = torch.randn(K, N, device=device, dtype=torch.float16)
>                 torch.cuda.synchronize()
>                 start_time = torch.cuda.Event(enable_timing=True)
>                 end_time = torch.cuda.Event(enable_timing=True)
> 
>                 start_time.record()
>                 C = torch.mm(A, B)
>                 end_time.record()
>                 torch.cuda.synchronize()
> 
>                 exec_time = start_time.elapsed_time(end_time) 
>                 total_times[round_idx] += exec_time
> 
>                 del A, B, C
>                 torch.cuda.empty_cache()
> 
>         print(f"for {round_idx+1} round time: {total_times[round_idx]:.2f}ms")
> 
>     print(f"- first round time (auto-tuning): {total_times[0]:.2f}ms")
> 
>     sum = 0
>     for _ in range(5):
>         sum += total_times[_+1] 
>     sum /= 5
>     print(f"- second round time (cached, with repeat five times): {sum:.2f}ms")
> 
> if __name__ == "__main__":
>     test_mm_configs()
> ```
> 
> the result as follow, disable early stop:
> 
> ```
> - first round time (auto-tuning): 503138.81ms
> - second round time (cached, with repeat five times): 167.11ms
> ```
> 
> enable early stop:
> 
> ```
> - first round time (auto-tuning): 252487.97ms
> - second round time (cached, with repeat five times): 175.43ms
> ```
> 
> Therefore, we can see that early stop can greatly reduce the cost of tuning, while having a relatively small impact on the performance of the operator.

What would happen with the increase of round number? If the selected config isn't the optimal, we may lose more than the benefits we gain from the auto-tuning. Will there be any mechanism to help decide when to enable early stopping?